requirements.txt
apscheduler
aiohttp
gevent
kafka-python
psycopg2
requests
pytest
pytest-aiohttp